on working

+++ database/migrations/2022_12_06_080901_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('product_nm', 255)->nullable()->comment('Product Name');
             $table->integer('category_id')->nullable()->comment('Category ID');
             $table->string('title', 255)->nullable()->comment('Title');
-            $table->float('price')->nullable()->comment('Price');
+            $table->decimal('price', 9, 2)->nullable()->comment('Price');
             $table->string('color', 255)->nullable()->comment('Color');
             $table->string('size', 20)->nullable()->comment('Size');
             $table->tinyInteger('is_new_product')->default(0)->comment('Is new product or not');
